Add files via upload
Created a URL shortener in C that compacts web URLs into short ones in the most effective way         possible. It is special for its exceptional hashing scheme and the absence of any considerable decorative elements in terms of its design; it is uniquely designed for high storage and swift retrieval at the same level.